work in progress, adding auth on queries.

+++ graphql/resolvers/accessGroupResolver.js
@@ -2,47 +2,58 @@ const AccessGroup = require("../../models/accessgroup/accessgroup");
 const Room = require("../../models/room/room");
 const User = require("../../models/user/user");
 const AccessGroupRoom = require("../../models/accessgrouproom/accessgrouproom");
-const AccessGroupUser = require("../../models/accessgroupuser/accessGroupUser");
+const AccessGroupUser = require("../../models/accessGroupUser/accessgroupuser");
 
 const resolver = {
   Query: {
-    accessGroups: async (parent, booking, { user }) => {
+    accessGroups: async (_, __, { user }) => {
       if (user.admin) {
         return AccessGroup.find({});
       }
       throw new Error("not authorized");
     },
-    accessGroup: async (_parent, args) =>
-      await AccessGroup.findById({ _id: args.id })
+    accessGroup: async (_, args, { user }) => {
+      if (user.admin) {
+        await AccessGroup.findById({ _id: args.id });
+      }
+      throw new Error("not authorized");
+    }
   },
 
   Mutation: {
-    addAccessGroup: async (_, args) => {
-      const newAccessGroup = await new AccessGroup(args).save();
-
-      return newAccessGroup;
+    addAccessGroup: async (_, args, { user }) => {
+      if (user.admin) {
+        return await new AccessGroup(args).save();
+      }
+      throw new Error("not authorized");
     },
-    removeAccessGroup: async (_, args) => {
-      AccessGroupRoom.deleteMany({ accessGroupId: args.id });
-      AccessGroupUser.deleteMany({ accessGroupId: args.id });
-
-      return await AccessGroup.findByIdAndDelete({ _id: args.id });
+    removeAccessGroup: async (_, args, { user }) => {
+      if (user.admin) {
+        await AccessGroupRoom.deleteMany({ accessGroupId: args.id });
+        await AccessGroupUser.deleteMany({ accessGroupId: args.id });
+        return await AccessGroup.findByIdAndDelete({ _id: args.id });
+      }
+      throw new Error("not authorized");
     },
-    updateAccessGroup: async (_, args) => {
-      return await AccessGroup.findOneAndUpdate({ _id: args.id }, { ...args });
+    updateAccessGroup: async (_, args, { user }) => {
+      if (user.admin) {
+        return await AccessGroup.findOneAndUpdate(
+          { _id: args.id },
+          { ...args }
+        );
+      }
+      throw new Error("not authorized");
     }
   },
   AccessGroup: {
     async rooms(parent) {
       let list = await AccessGroupRoom.find({ accessGroupId: parent._id });
-
       return Room.find({
         _id: { $in: list.map(item => item.roomId) }
       });
     },
     async users(parent) {
       let list = await AccessGroupUser.find({ accessGroupId: parent._id });
-
       return User.find({ _id: { $in: list.map(item => item.userId) } });
     }
   }
